Catch ConnectClosedError from junos-eznc
And raise ConnectionClosedException from napalm-base.
We can easily raise what we need using a metaclass that
catches the exceptions and raises the desired exception as required.
This models has the advantage that it's easier to raise other kind
of exceptions. We will soon need to raise other new exceptions
from napalm-base, e.g.: LockException, UnlockException etc.